refactor: Refactored some tests, consolidated main Python modules
