+++ run_test_app.py
@@ -8,7 +8,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).parent / "src"))
 
 if __name__ == "__main__":
-    from accessiweather.simple.test_app import main
+    from accessiweather.test_app import main
 
     app = main()
     app.main_loop()
