+++ src/accessiweather/__main__.py
@@ -3,7 +3,7 @@
 Allows running the module with `python -m accessiweather`
 """
 
-from accessiweather.toga_app import main
+from accessiweather.app import main
 
 if __name__ == "__main__":
     main().main_loop()
